# openapi/spec.yaml
openapi: "3.0.3"
info:
  title: ADH6 API
  description: >
    This is the specification for **MiNET**'s ADH6 plaform. Its aim is to manage our users, devices and treasury.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 2.0.0
  contact:
    name: MiNET
    email: equipe@minet.net
    url: https://minet.net
servers:
  - url: https://adh6.minet.net/api
    description: ADH6 production server
  - url: https://adh6-dev.minet.net/api
    description: ADH6 staging server
  - url: https://adh6-local.minet.net/api
    description: ADH6 staging server
paths:
...
  /profile:
    get:
      summary: Introspection endpoint
      description: >
        This endpoint returns information about the currently logged-in user.
      operationId: profile
      responses:
        "200":
          description: The current logged-in user
          content:
            application/json:
              schema:
                properties:
                  member:
                    $ref: "#/components/schemas/Member"
                  roles:
                    type: array
                    items:
                      type: string
        "403":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: '#/components/responses/Unexpected'      
      security: 
        - OpenIdConnect:
          - user
        - OpenIdConnect:
          - admin:read
      tags:
      - misc
...
components:
...
  responses:
...
    Unauthorized:
      description: The user has insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unexpected:
      description: The request encountered an unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    AbstractMember:
      type: object
      properties:
        id:
          description: The unique identifier of this member
          example: 0
          readOnly: true
          type: integer
        username:
          description: The username of this member
          example: michel_j
          type: string
        firstName:
          description: The first name of this member
          example: Jean-Michel
          type: string
        lastName:
          description: The last name of this member
          example: Martin
          type: string
        email:
          description: The email address of this member
          example: jean-michel.martin@example.com
          type: string
          format: email
        comment:
          description: The optional comment for this member
          nullable: true
          type: string
          example: Aime bien les bananes
        departureDate:
          description: The date-time at which the subscription of this member expires
          format: date-time
          type: string
        mailinglist:
          deprecated: true
          description: The mailing lists the member is subscribed to (bit field)
          type: integer
          example: 4
        ip:
          description: The public ipv4 address assigned to this member
          example: 157.159.192.144
          nullable: true
          format: ipv4
          type: string
        subnet:
          description: The private ipv4 network assigned to this member
          example: 10.42.8.240/28
          nullable: true
          type: string
        membership:
          type: string
          nullable: true
          enum:
            - INITIAL
            - PENDING_RULES
            - PENDING_PAYMENT_INITIAL
            - PENDING_PAYMENT
            - PENDING_PAYMENT_VALIDATION
            - COMPLETE
            - CANCELLED
            - ABORTED
          example: INITIAL
          description: >
            The current status of this membership request:
             * `INITIAL` - Just created
             * `PENDING_RULES` - Waiting for the member to sign the rules
             * `PENDING_PAYMENT_INITIAL` - Initiating the payment flow
             * `PENDING_PAYMENT` - During the payment flow
             * `PENDING_PAYMENT_VALIDATION` - After the payment flow, waiting for confirmation
             * `COMPLETE` - The membership request is completed
             * `CANCELLED` - The membership has been cancelled
             * `ABORTED` - The membership request flow was aborted
            Do note that some of the steps may be skipped depending on the payment method,
            whether or not this is the member's first membership request etc.
    Member:
      allOf:
        - $ref: "#/components/schemas/AbstractMember"
        - required:
            - id
            - username
            - email
            - firstName
            - lastName
      type: object
...
    Error:
      type: object
      properties:
        code:
          description: The error code
          type: integer
          example: 404
        message:
          description: The friendly description of the error
          type: string
          example: Not found
...
  securitySchemes:
    OpenIdConnect:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/auth
          tokenUrl: https://keycloak.minet.net/auth/realms/MiNET/protocol/openid-connect/token
          scopes:
            user: Basic user access
            admin:read: Read access to admin resources
            admin:write: Write access to admin resources
            admin:prod: Production admin access
            treasurer:read: Read access to treasury resources
            treasurer:write: Write access to treasury resources
            network:read: Read access to network resources
            network:write: Write access to network resources
            network:write:prod: Production network write access
            network:write:dev: Development network write access
            network:write:hosting: Hosting network write access
    ApiKeyAdminAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_admin_auth
    ApiKeyNetworkAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_network_auth
    ApiKeyTreasuryAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      x-apikeyInfoFunc: adh6.authentication.apikey_treasury_auth
...